Refactor to state pattern for FTPClient
