+++ src/main/java/parsing/WriteCommand.java
@@ -1,28 +1,16 @@
 package parsing;
 
-import exceptions.InvalidPacketException;
-import formats.AckMessage;
-import formats.DataMessage;
-import formats.Message.MessageType;
-import formats.RequestMessage;
 import logging.Logger;
-import resources.ResourceManager;
-import socket.TFTPDatagramSocket;
+import states.InputState;
+import states.State;
+import states.WriteState;
 
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.net.DatagramPacket;
-import java.net.SocketAddress;
-import java.net.SocketException;
 import java.net.UnknownHostException;
 import java.util.List;
-
-import static resources.Configuration.GLOBAL_CONFIG;
 
 public class WriteCommand extends SocketCommand {
     private static final Logger LOG = new Logger("FTPClient - Write");
 	private static final String WRITE_OPERATION = "write";
-	private TFTPDatagramSocket socket;
 
 	public WriteCommand(List<String> tokens) {
 		super(WRITE_OPERATION, tokens);
@@ -31,91 +19,17 @@ public class WriteCommand extends SocketCommand {
 	public WriteCommand(String[] tokens) {
 		super(WRITE_OPERATION, tokens);
 	}
-
+	
 	@Override
-	public void execute() {
-        if(this.tokens.size() < 3) {
-            LOG.logQuiet("Error: Not enough arguments");
-        }
-        else {
-            try {
-                ResourceManager resourceManager = new ResourceManager(GLOBAL_CONFIG.CLIENT_RESOURCE_DIR);
-                if(!resourceManager.fileExists(this.getFilename()))
-                    throw new FileNotFoundException("File Not Found");
-
-                socket = new TFTPDatagramSocket();
-                socket.setSoTimeout(SOCKET_TIMEOUT);
-
-                try {
-
-                    // Create the write request & send
-                    RequestMessage wrqMessage = new RequestMessage(MessageType.WRQ ,this.getFilename());
-                    socket.sendMessage(wrqMessage, this.getServerAddress());
-                    LOG.logQuiet("---- Begin File Transaction ---");
-                    LOG.logQuiet("Write Request has been sent!");
-
-
-                    // Wait for WRQ ack
-                    DatagramPacket recv = socket.receivePacket();
-                    AckMessage ack = AckMessage.parseMessageFromPacket(recv);
-                    LOG.logVerbose("Received WRQ ACK");
-
-                    if(ack.getBlockNum() != 0) throw new IOException("Incorrect Initial Block Number");
-
-                    sendDataBlock(resourceManager.readFileToBytes(this.getFilename()), recv.getSocketAddress());
-                    LOG.logQuiet("Successfully completed write operation");
-                    LOG.logQuiet("---- End File Transaction ---");
-                } catch(InvalidPacketException ipE) {
-                    LOG.logQuiet(ipE.getMessage());
-                } catch(UnknownHostException uHE) {
-                    LOG.logQuiet("Error: Unknown Host Entered");
-                } catch(IOException ioE) {
-                    ioE.printStackTrace();
-                }
-                socket.close();
-            } catch (SocketException sE) {
-                    sE.printStackTrace();
-            } catch (FileNotFoundException fNFE) {
-                LOG.logQuiet("Error: " + fNFE.getMessage());
-            } catch (IOException ioE) {
-                LOG.logVerbose("IOException occurred. " + ioE.getLocalizedMessage());
-            }
-        }
+	public State execute() {
+		if (this.tokens.size() < 3)
+			LOG.logQuiet("Error: Not enough arguments");
+		else
+			try {
+				return new WriteState(getServerAddress(), getFilename(), isVerbose());
+			} catch (UnknownHostException e) {
+				e.printStackTrace();
+			}
+		return new InputState();
 	}
-
-	/**
-	 * Sends an array of bytes over TFTP, splicing the data in blocks if necessary
-     * @param data The array of data
-     * @param socketAddress The Socket address used in sending packet
-     * @throws IOException
-    */
-    private void sendDataBlock(byte[] data, SocketAddress socketAddress) throws IOException {
-
-        // Create a sequence of data for the byte array
-        List<DataMessage> messages = DataMessage.createDataMessageSequence(data);
-        LOG.logVerbose("Sending " + messages.size() + " Data messages to " + socketAddress);
-
-        try {
-            // Send each data block sequentially
-            for(DataMessage m : messages) {
-
-                LOG.logVerbose("Block: " + m.getBlockNum() + ", Data size: " + m.getDataSize() + ", Final Block: " + m.isFinalBlock());
-
-                // Send Data block message
-                socket.sendMessage(m, socketAddress);
-
-                // Wait for Ack message before continuing
-                DatagramPacket packet = socket.receivePacket();
-                AckMessage ack = AckMessage.parseMessageFromPacket(packet);
-                LOG.logVerbose("Ack Received: " + ack.getBlockNum());
-
-                if(ack.getBlockNum() != m.getBlockNum()) {
-                    throw new InvalidPacketException("Invalid Block Number");
-                }
-            }
-        } catch (InvalidPacketException e) {
-            e.printStackTrace();
-            throw new IOException("Failed to parse Ack Message", e);
-        }
-    }
 }
