+++ src/main/java/states/ExitState.java
@@ -1,0 +1,11 @@
+package states;
+
+public class ExitState extends State {
+
+	@Override
+	public State execute() {
+		System.exit(0);
+		return null;
+	}
+
+}
