+++ src/main/java/states/InputState.java
@@ -1,0 +1,19 @@
+package states;
+
+import java.util.Scanner;
+
+import parsing.Command;
+import parsing.Parser;
+
+public class InputState extends State {
+	@Override
+	public State execute() {
+		System.out.println("Type 'help' for a list of commands");
+		
+		Scanner sc = new Scanner(System.in);
+		String input = sc.nextLine();
+		sc.close();
+		Command command = Parser.parse(input.split(" "));
+		return command.execute();
+	}
+}
