+++ src/main/java/states/State.java
@@ -1,0 +1,10 @@
+package states;
+
+import logging.Logger;
+
+
+public abstract class State {
+	protected static final Logger LOG = new Logger("FTPClient");
+		
+	public abstract State execute();
+}
